+++ frontend/office/src/pages/RecipeDetailPage.tsx
@@ -1,13 +1,13 @@
-import { useEffect, useState } from 'react';
 import { useParams, useNavigate } from 'react-router-dom';
 import { useDeleteRecipe, useRecipe, useUpdateRecipe } from '@/hooks/useRecipes';
 import { Button } from '@/components/ui/button';
-import { Input } from '@/components/ui/input';
-import { Textarea } from '@/components/ui/textarea';
-import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card';
+import { InlineEditableRecipeText } from '../components/recipes/InlineEditableRecipeText';
+import { Section } from '../components/Section';
+import { formatDatetime } from '../components/Format-Datetime.tsx';
+import { DetailRow } from '../components/Detail-row';
+import { Grid } from '../components/Grid';
 import { cn } from '@/lib/utils';
 import recipeImage from '@/assets/voice-chef-recipe.jpg';
-import type { Recipe } from '@/types/recipe';
 
 const STATUS_STYLES: Record<string, string> = {
   draft: 'bg-yellow-100 text-yellow-800',
@@ -15,172 +15,6 @@ const STATUS_STYLES: Record<string, string> = {
   archived: 'bg-gray-100 text-gray-600',
 };
 
-function DetailRow({ label, value }: { label: string; value: string | number | boolean | null | undefined }) {
-  const isEmpty = value == null || value === '';
-  const display = isEmpty
-    ? null
-    : typeof value === 'boolean'
-    ? value ? 'Yes' : 'No'
-    : String(value);
-  return (
-    <div className="flex flex-col gap-0.5">
-      <span className="text-xs text-muted-foreground uppercase tracking-wide">{label}</span>
-      {isEmpty
-        ? <span className="text-sm text-muted-foreground/50 italic">empty</span>
-        : <span className="text-sm font-medium">{display}</span>
-      }
-    </div>
-  );
-}
-
-function Section({ title, children }: { title: string; children: React.ReactNode }) {
-  return (
-    <Card>
-      <CardHeader className="pb-2">
-        <CardTitle className="text-sm uppercase tracking-wide text-muted-foreground font-medium">
-          {title}
-        </CardTitle>
-      </CardHeader>
-      <CardContent>{children}</CardContent>
-    </Card>
-  );
-}
-
-function Grid({ children }: { children: React.ReactNode }) {
-  return <div className="grid grid-cols-2 sm:grid-cols-3 gap-x-6 gap-y-4">{children}</div>;
-}
-
-type EditableRecipeField = 'name' | 'description' | 'instructions';
-
-function InlineEditableText({
-  recipeId,
-  field,
-  value,
-  label,
-  multiline = false,
-  className,
-  displayClassName,
-}: {
-  recipeId: string;
-  field: EditableRecipeField;
-  value: string | null | undefined;
-  label: string;
-  multiline?: boolean;
-  className?: string;
-  displayClassName?: string;
-}) {
-  const updateRecipe = useUpdateRecipe();
-  const [isEditing, setIsEditing] = useState(false);
-  const [draft, setDraft] = useState(value ?? '');
-  const [validationError, setValidationError] = useState<string | null>(null);
-
-  useEffect(() => {
-    if (!isEditing) {
-      setDraft(value ?? '');
-      setValidationError(null);
-    }
-  }, [isEditing, value]);
-
-  const handleSave = () => {
-    const nextValue = draft.trim();
-
-    if (field === 'name' && !nextValue) {
-      setValidationError('Recipe name is required.');
-      return;
-    }
-
-    setValidationError(null);
-
-    const payload =
-      field === 'name'
-        ? ({ id: recipeId, name: nextValue } as Partial<Recipe> & { id: string })
-        : ({ id: recipeId, [field]: nextValue || null } as Partial<Recipe> & { id: string });
-
-    updateRecipe.mutate(
-      payload,
-      {
-        onSuccess: () => {
-          setValidationError(null);
-          setIsEditing(false);
-        },
-      },
-    );
-  };
-
-  return (
-    <div className={cn('space-y-1', className)}>
-      {label && <span className="text-xs text-muted-foreground uppercase tracking-wide">{label}</span>}
-      {isEditing ? (
-        <div className="space-y-2">
-          {multiline ? (
-            <Textarea
-              value={draft}
-              onChange={(event) => {
-                setDraft(event.target.value);
-                if (validationError) {
-                  setValidationError(null);
-                }
-              }}
-              autoFocus
-            />
-          ) : (
-            <Input
-              value={draft}
-              className={cn(field === 'name' && 'h-12 text-2xl font-semibold')}
-              onChange={(event) => {
-                setDraft(event.target.value);
-                if (validationError) {
-                  setValidationError(null);
-                }
-              }}
-              autoFocus
-            />
-          )}
-          {validationError && (
-            <p className="text-xs text-destructive">{validationError}</p>
-          )}
-          <div className="flex gap-2">
-            <Button size="sm" type="button" onClick={handleSave} disabled={updateRecipe.isPending}>
-              Save
-            </Button>
-            <Button
-              size="sm"
-              variant="outline"
-              type="button"
-              onClick={() => {
-                setDraft(value ?? '');
-                setValidationError(null);
-                setIsEditing(false);
-              }}
-            >
-              Cancel
-            </Button>
-          </div>
-        </div>
-      ) : value ? (
-        <button
-          type="button"
-          onClick={() => setIsEditing(true)}
-          className="block w-full text-left rounded-lg border border-transparent px-2 py-1 -mx-2 -my-1 hover:border-border hover:bg-muted/40 transition-colors"
-        >
-          {multiline ? (
-            <p className={cn('text-sm leading-relaxed whitespace-pre-line', displayClassName)}>{value}</p>
-          ) : (
-            <p className={cn('text-sm font-medium', displayClassName)}>{value}</p>
-          )}
-        </button>
-      ) : (
-        <button
-          type="button"
-          onClick={() => setIsEditing(true)}
-          className="w-full text-left rounded-lg border border-dashed border-border/70 px-2 py-1 text-sm text-muted-foreground/50 italic hover:bg-muted/30 transition-colors"
-        >
-          Click to add
-        </button>
-      )}
-    </div>
-  );
-}
 
 // function TextBlock({ label, value }: { label: string; value: string | null | undefined }) {
 //   return (
@@ -198,11 +32,6 @@ function InlineEditableText({
 //   if (!value) return null;
 //   return new Date(value).toLocaleDateString(undefined, { year: 'numeric', month: 'short', day: 'numeric' });
 // }
-
-function formatDatetime(value: string | null | undefined) {
-  if (!value) return null;
-  return new Date(value).toLocaleString(undefined, { year: 'numeric', month: 'short', day: 'numeric', hour: '2-digit', minute: '2-digit' });
-}
 
 
 export function RecipeDetailPage() {
@@ -292,7 +121,7 @@ export function RecipeDetailPage() {
           )}
         </div>
         <div className="flex items-center gap-3 flex-wrap">
-          <InlineEditableText
+          <InlineEditableRecipeText
             recipeId={recipe.id}
             field="name"
             value={recipe.name}
@@ -331,9 +160,63 @@ export function RecipeDetailPage() {
         </div>
       </div>
 
+      {/* <Section title="preparation_time_minutes">
+        preparation_time_minutes
+      </Section> */}
+
+      {/* ── 4 cols: preparation_time_minutes ── */}
+      <div className="grid grid-cols-1 lg:grid-cols-12 gap-10 items-start">
+
+      {/* 1 column */}
+      <div className="lg:col-span-3">
+       <Section title="PREP TIME">
+        {recipe.preparation_time_minutes}
+        {(!recipe.preparation_time_minutes || recipe.preparation_time_minutes === 0) && (
+          <p className="text-sm text-muted-foreground italic">No preparation time minutes added yet.</p>
+          )}
+       </Section>
+      </div>
+
+      {/* column 2 */}
+      <div className="lg:col-span-3">
+       <Section title="COOK TIME">
+        {recipe.cooking_time_minutes}
+        {(!recipe.cooking_time_minutes || recipe.cooking_time_minutes === 0) && (
+          <p className="text-sm text-muted-foreground italic">No cooking time minutes added yet.</p>
+          )}
+       </Section>
+      </div>
+      {/* column 3 */}
+      <div className="lg:col-span-3">
+       <Section title="SERVINGS">
+        {/* Ternary operator for conditional rendering to ensure the math 
+        only happens if a valid value exists. */}
+        {recipe.portions_count_resolved ? (
+          // Only render the number if portion_size_grams is not null or empty
+          Math.round(Number(recipe.portions_count_resolved))
+        ) : (
+          // Fallback message if it is null, 0, or an empty string
+          <p className="text-sm text-muted-foreground italic">No portions count resolved added yet.</p>
+        )}
+       </Section>
+      </div>
+      {/* column 4 */}
+      <div className="lg:col-span-3">
+       <Section title="PORTION SIZE">
+        {recipe.portion_size_grams ? (
+          Math.round(Number(recipe.portion_size_grams))
+        ) : (
+          <p className="text-sm text-muted-foreground italic">
+            No portion size grams resolved added yet.
+          </p>
+        )}
+       </Section>
+      </div>
+    </div>
+
       {/* ── Long-form text ─────────────────────────────────── */}
       <Section title="Description">
-        <InlineEditableText
+        <InlineEditableRecipeText
           recipeId={recipe.id}
           field="description"
           value={recipe.description}
@@ -376,7 +259,7 @@ export function RecipeDetailPage() {
       <div className="lg:col-span-8">
        <Section title="Instructions">
         <div className="min-h-[200px]">
-          <InlineEditableText
+          <InlineEditableRecipeText
             recipeId={recipe.id}
             field="instructions"
             value={recipe.instructions}
@@ -385,8 +268,9 @@ export function RecipeDetailPage() {
           />
         </div>
        </Section>
-     </div>
-   </div>
+      </div>
+    </div>
+
          {/* ── Identity ───────────────────────────────────────── */}
       <Section title="Identity">
         <Grid>
@@ -402,14 +286,8 @@ export function RecipeDetailPage() {
       {/* <Section title="total_raw_weight_grams">
         total_raw_weight_grams
       </Section> */}
-      {/* <Section title="portions_count_resolved">
-        portions_count_resolved
-      </Section> */}
       {/* <Section title="photo_url">
         photo_url
-      </Section> */}
-      {/* <Section title="preparation_time_minutes">
-        preparation_time_minutes
       </Section> */}
       {/* ── Yield & Weights ────────────────────────────────── */}
       <Section title="Yield & Weights">
@@ -417,7 +295,6 @@ export function RecipeDetailPage() {
             {/* <DetailRow label="Yield" value={recipe.yield_amount != null ? `${recipe.yield_amount}${recipe.yield_unit ? ` ${recipe.yield_unit}` : ''}` : null} /> */}
             {/* <DetailRow label="Reduction factor" value={recipe.reduction_factor} /> */}
             <DetailRow label="Total cooked weight (g)" value={recipe.total_cooked_weight_grams} />
-            <DetailRow label="Portion size (g)" value={recipe.portion_size_grams} />
           </Grid>
         </Section>
     </div>
